style: changes and formattings
