+++ client/package.json
@@ -9,6 +9,7 @@
     "pubnub": "^4.28.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
+    "react-icons": "^3.10.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.1"
   },
