+++ client/src/components/LinkDisplay/style.css
@@ -1,0 +1,36 @@
+.container {
+  background: #fff;
+  box-shadow: 0 5px 5px rgba(0, 0, 0, 0.05), 0 6px 6px rgba(0, 0, 0, 0.05);
+  color: #fff;
+  display: flex;
+  padding: 2px;
+  flex: 1;
+  border-radius: 5px;
+  margin-top: 20px;
+  align-items: center;
+  justify-content: center;
+  text-align: center;
+  padding-left: 10px;
+}
+
+.linkArea {
+  width: 95%;
+  border-width: 0px;
+  font-family: inherit;
+  font-size: 18px;
+  outline: None;
+  padding-top: 10px;
+  height: 25px;
+}
+
+.copyButton {
+  border-width: 0.5px;
+  border-color: #f0f0f0;
+  font-family: inherit;
+  font-size: 18px;
+  width: 100px;
+  height: 40px;
+  border-radius: 3px;
+  cursor: pointer;
+  outline: none;
+}
