+++ package.json
@@ -22,7 +22,6 @@
     "mongodb": "^3.5.9",
     "node-fetch": "^2.6.0",
     "pubnub": "^4.28.2",
-    "react-icons": "^3.10.0",
     "uuid": "^8.2.0"
   },
   "devDependencies": {
